# Conflicts:
#	src/main/java/com/css/autocsfinal/market/repository/StoreInfo2Repository.java
#	src/main/java/com/css/autocsfinal/member/controller/MemberController.java
#	src/main/java/com/css/autocsfinal/member/service/MemberService.java

+++ src/main/java/com/css/autocsfinal/stock/controller/IoController.java
@@ -81,19 +81,26 @@ public class IoController {
 
         List<IoSummaryDTO> ioSummaryPage = tuplePage.stream()
                 .map(tuple -> {
-                    int refProductNo = tuple.get(0, Integer.class);
-                    long totalQuantityIn = tuple.get(1, Long.class);
-                    long totalQuantityOut = tuple.get(2, Long.class);
+                    BigDecimal refProductNo =  tuple.get(0, BigDecimal.class);
+                    BigDecimal totalQuantityIn = tuple.get(1, BigDecimal.class);
+                    BigDecimal totalQuantityOut = tuple.get(2, BigDecimal.class);
+//                    int refProductNo = tuple.get(0, Integer.class);
+//                    long totalQuantityIn = tuple.get(1, Long.class);
+//                    long totalQuantityOut = tuple.get(2, Long.class);
                     String productName = tuple.get(3, String.class);
                     String categoryName = tuple.get(4, String.class);
                     String standardName = tuple.get(5, String.class);
                     String unitName = tuple.get(6, String.class);
-                    int stock = tuple.get(7, Integer.class);
-                    int price = tuple.get(8, Integer.class);
+                    BigDecimal stock = tuple.get(7, BigDecimal.class);
+                    BigDecimal price = tuple.get(8, BigDecimal.class);
+//                    int stock = tuple.get(7, Integer.class);
+//                    int price = tuple.get(8, Integer.class);
                     String etc = tuple.get(9, String.class);
 
-                    return new IoSummaryDTO(refProductNo, totalQuantityIn, totalQuantityOut, productName, categoryName, standardName,
-                            unitName, stock, price, etc );
+                    return new IoSummaryDTO(refProductNo.intValue(), totalQuantityIn.intValue(), totalQuantityOut.intValue(), productName, categoryName, standardName,
+                            unitName, stock.intValue(), price.intValue(), etc );
+//                    return new IoSummaryDTO(refProductNo, totalQuantityIn, totalQuantityOut, productName, categoryName, standardName,
+//                            unitName, stock, price, etc );
                 })
                 .collect(Collectors.toList());
